Added script to show reproducibility. (#17)
* Added script to show reproducibility.

* Added instructions for verify reproducibility script

* Incorporated PR comments (Removed unused imports and naming)

+++ simulator/opendc-experiments/opendc-experiments-allocateam/tools/run.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
+# This script runs the Allocateam experiment.
+# Usage: ./tools/run.sh smokeTest
 
 set -euo pipefail
 
@@ -14,12 +16,19 @@ command -v realpath > /dev/null 2>&1 || realpath() {
 
 # Check if data directory exists and ask user what to do as we don't want to remove data
 if [ -d "$DATA_DIR" ]; then
-	read -rp "data directory exists. Remove it (y/n)? " choice
-	case "$choice" in
-		y|Y ) rm -r "$DATA_DIR" && mkdir "$DATA_DIR";;
-		n|N ) echo "Cannot proceed without (re)moving data directory. Aborting..."; exit 1;;
-		* ) echo "Ambiguous answer. Aborting..."; exit 2;;
-	esac
+	while getopts ":y" opt; do # if -y option is given, remove data dir.
+		case $opt in
+			y) echo "Removing data directory.." && rm -r "$DATA_DIR" && mkdir "$DATA_DIR";;
+		esac
+		done
+	if [ $OPTIND -eq 1 ]; then  # if no options given, prompt for data dir removal.
+		read -rp "data directory exists. Remove it (y/n)? " choice
+		case "$choice" in
+			y|Y ) rm -r "$DATA_DIR" && mkdir "$DATA_DIR";;
+			n|N ) echo "Cannot proceed without (re)moving data directory. Aborting..."; exit 1;;
+			* ) echo "Ambiguous answer. Aborting..."; exit 2;;
+		esac
+	fi
 else
 	mkdir "$DATA_DIR"
 fi
